+++ fugitive-darkness-api/src/main/java/com/zer0s2m/fugitivedarkness/api/handlers/ControllerApiGitRepoInstall.java
@@ -66,7 +66,7 @@ final public class ControllerApiGitRepoInstall implements Handler<RoutingContext
                 ))
                 .onComplete(resultSaved -> {
                     if (!resultSaved.succeeded()) {
-                        logger.error("Failure: " + resultSaved.cause());
+                        logger.error("Failure (DB 1): " + resultSaved.cause());
                         return;
                     }
 
@@ -80,9 +80,11 @@ final public class ControllerApiGitRepoInstall implements Handler<RoutingContext
                     executor
                             .executeBlocking(() -> {
                                 try {
-                                    return serviceGit.gClone(containerGitRepoInstall.remote());
+                                    return serviceGit.gCloneStart(
+                                            serviceGit.gCloneCreate(containerGitRepoInstall.remote()),
+                                            containerGitRepoInstall.remote());
                                 } catch (GitAPIException e) {
-                                    logger.error("Failure (GIT): " + e.fillInStackTrace());
+                                    logger.error("Failure (GIT 1): " + e.fillInStackTrace());
                                     throw new RuntimeException(e);
                                 }
                             }, false)
@@ -90,14 +92,14 @@ final public class ControllerApiGitRepoInstall implements Handler<RoutingContext
                                     .deleteByGroupAndProject(infoRepo.group(), infoRepo.project())
                                     .onSuccess(ar -> repositoryGit.closeClient())
                                     .onFailure(fail2 -> {
-                                        logger.error("Failure (DB 1): " + fail2.getMessage());
+                                        logger.error("Failure (DB 2): " + fail2.getMessage());
                                         repositoryGit.closeClient();
                                     }))
                             .onSuccess(result -> repositoryGit
                                     .updateIsLoadByGroupAndProject(result.group(), result.project(), true)
                                     .onComplete(ar -> {
                                         if (!ar.succeeded()) {
-                                            logger.error("Failure (DB 2): " + ar.cause());
+                                            logger.error("Failure (GIT 2): " + ar.cause());
                                         }
 
                                         repositoryGit.closeClient();