+++ fugitive-darkness-common/src/main/java/com/zer0s2m/fugitivedarkness/common/dto/ContainerInfoGitProviderControl.java
@@ -1,6 +1,6 @@
 package com.zer0s2m.fugitivedarkness.common.dto;
 
-public record ContainerInfoGitProviderDelete(
+public record ContainerInfoGitProviderControl(
         String type,
 
         String target