+++ fugitive-darkness-provider/src/main/java/com/zer0s2m/fugitivedarkness/provider/RepositoryInstallerRemote.java
@@ -1,0 +1,28 @@
+package com.zer0s2m.fugitivedarkness.provider;
+
+import com.zer0s2m.fugitivedarkness.provider.impl.GitRepoImpl;
+import org.eclipse.jgit.api.errors.GitAPIException;
+
+public interface RepositoryInstallerRemote {
+
+    /**
+     * Collect runtime code to clone the repository.
+     *
+     * @param URI Remote git repository host. Must not be {@literal null}.
+     * @return Execution code.
+     */
+    static Runnable taskInstallRepoRemote(final String URI) {
+        return () -> {
+            final GitRepo gitRepoService = new GitRepoImpl();
+            try {
+                gitRepoService.gCloneStart(
+                        gitRepoService.gCloneCreate(URI),
+                        URI
+                );
+            } catch (GitAPIException e) {
+                throw new RuntimeException(e);
+            }
+        };
+    }
+
+}